Progress on a .NET Core build
This does the following:
- Adds netstandard builds (almost, read on)
- Moves to the new .csproj system
- Removes the dependency on the Windows SDK install
- Adds an overall build script (currently a demo in failure!)

ILASM is throwing, but most of the infrastructure to make this work is in place. There's a fundamental bootstrapping issue with the need to publish ConstraintChanger to get ILASM and ILDASM out of NuGet package cleanly, so a one-time "build.bat" run will be needed for new users.

Beyond that, ILASM goes boom - checkout and run build.bat to see everything in the current state.

Notes: netstandard doesn't have CreateDelegate() too early, and I'm unsure if tests pass. If we get ilasm working tests will probably be in good shape. Also, while it should generate a cross-platform assembly, it will only *build* on Windows (we have to pick something to get ilasm/ildasm). The resulting package should be consumable anywhere.